+++ packages/stitch/src/index.ts
@@ -1,6 +1,5 @@
 export { stitchSchemas } from './stitchSchemas';
 export { createMergedTypeResolver } from './createMergedTypeResolver';
 export { forwardArgsToSelectionSet } from './selectionSetArgs';
-export { isolateComputedFields } from './isolateComputedFields';
 
 export * from './subschemaConfigTransforms';