+++ packages/stitching-directives/src/properties.ts
@@ -1,6 +1,6 @@
 import { PropertyTree } from './types';
 
-export function addKey(object: Record<string, any>, path: Array<string | number>, value: any) {
+export function addProperty(object: Record<string, any>, path: Array<string | number>, value: any) {
   const initialSegment = path[0];
   if (path.length === 1) {
     object[initialSegment] = value;
@@ -9,7 +9,7 @@ export function addKey(object: Record<string, any>, path: Array<string | number>
 
   let field = object[initialSegment];
   if (field != null) {
-    addKey(field, path.slice(1), value);
+    addProperty(field, path.slice(1), value);
     return;
   }
 
@@ -18,30 +18,55 @@ export function addKey(object: Record<string, any>, path: Array<string | number>
   } else {
     field = [];
   }
-  addKey(field, path.slice(1), value);
+  addProperty(field, path.slice(1), value);
   object[initialSegment] = field;
 }
 
-export function getKey(object: Record<string, any>, path: Array<string>): any {
-  return path.reduce((acc, pathSegment) => acc[pathSegment], object);
+export function getProperty(object: Record<string, any>, path: Array<string>): any {
+  if (!path.length || object == null) {
+    return object;
+  }
+
+  const newPath = path.slice();
+  const key = newPath.shift();
+  const prop = object[key];
+
+  return getProperty(prop, newPath);
 }
 
-export function getKeys(object: Record<string, any>, propertyTree: PropertyTree): any {
+export function getProperties(object: Record<string, any>, propertyTree: PropertyTree): any {
+  if (object == null) {
+    return object;
+  }
+
   const newObject = Object.create(null);
+
   Object.entries(propertyTree).forEach(([key, subKey]) => {
     if (subKey == null) {
       newObject[key] = object[key];
-    } else {
-      newObject[key] = getKeys(object[key], subKey);
+      return;
     }
+
+    const prop = object[key];
+
+    newObject[key] = deepMap(prop, (item) => getProperties(item, subKey));
   });
+
   return newObject;
 }
 
 export function propertyTreeFromPaths(paths: Array<Array<string>>): PropertyTree {
   const propertyTree = Object.create(null);
   paths.forEach(path => {
-    addKey(propertyTree, path, null);
+    addProperty(propertyTree, path, null);
   });
   return propertyTree;
+}
+
+function deepMap(arrayOrItem: any, fn: (item: any) => any): any {
+  if (Array.isArray(arrayOrItem)) {
+    return arrayOrItem.map(nestedArrayOrItem => deepMap(nestedArrayOrItem, fn));
+  }
+
+  return fn(arrayOrItem);
 }